// packages/protocol/typescript/generated/yiru/runtime/v1/git_remote_pb.d.ts
export declare type GitRemoteServiceForkSyncResponse = Message<"yiru.runtime.v1.GitRemoteServiceForkSyncResponse"> & {
  /**
   * @generated from field: string origin_remote = 1;
   */
  originRemote: string;

  /**
   * @generated from field: string upstream_remote = 2;
   */
  upstreamRemote: string;

  /**
   * @generated from field: uint64 ahead = 3;
   */
  ahead: bigint;

  /**
   * @generated from field: uint64 behind = 4;
   */
  behind: bigint;

  /**
   * @generated from field: yiru.runtime.v1.GitForkSyncStatus status = 5;
   */
  status: GitForkSyncStatus;

  /**
   * @generated from field: optional yiru.runtime.v1.GitForkSyncBlockReason reason = 6;
   */
  reason?: GitForkSyncBlockReason | undefined;

  /**
   * @generated from field: optional string branch_name = 7;
   */
  branchName?: string | undefined;
};

/**
 * Describes the message yiru.runtime.v1.GitRemoteServiceForkSyncResponse.
 * Use `create(GitRemoteServiceForkSyncResponseSchema)` to create a new message.
 */
export declare const GitRemoteServiceForkSyncResponseSchema: GenMessage<GitRemoteServiceForkSyncResponse>;

/**
 * @generated from enum yiru.runtime.v1.GitForkSyncStatus
 */
export enum GitForkSyncStatus {
  /**
   * @generated from enum value: GIT_FORK_SYNC_STATUS_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: GIT_FORK_SYNC_STATUS_BLOCKED = 1;
   */
  BLOCKED = 1,

  /**
   * @generated from enum value: GIT_FORK_SYNC_STATUS_UP_TO_DATE = 2;
   */
  UP_TO_DATE = 2,

  /**
   * @generated from enum value: GIT_FORK_SYNC_STATUS_SYNCED = 3;
   */
  SYNCED = 3,
}

/**
 * Describes the enum yiru.runtime.v1.GitForkSyncStatus.
 */
export declare const GitForkSyncStatusSchema: GenEnum<GitForkSyncStatus>;

/**
 * @generated from enum yiru.runtime.v1.GitForkSyncBlockReason
 */
export enum GitForkSyncBlockReason {
  /**
   * @generated from enum value: GIT_FORK_SYNC_BLOCK_REASON_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: GIT_FORK_SYNC_BLOCK_REASON_MISSING_ORIGIN = 1;
   */
  MISSING_ORIGIN = 1,

  /**
   * @generated from enum value: GIT_FORK_SYNC_BLOCK_REASON_MISSING_UPSTREAM = 2;
   */
  MISSING_UPSTREAM = 2,

  /**
   * @generated from enum value: GIT_FORK_SYNC_BLOCK_REASON_UPSTREAM_MISMATCH = 3;
   */
  UPSTREAM_MISMATCH = 3,

  /**
   * @generated from enum value: GIT_FORK_SYNC_BLOCK_REASON_MISSING_UPSTREAM_DEFAULT_BRANCH = 4;
   */
  MISSING_UPSTREAM_DEFAULT_BRANCH = 4,

  /**
   * @generated from enum value: GIT_FORK_SYNC_BLOCK_REASON_MISSING_ORIGIN_BRANCH = 5;
   */
  MISSING_ORIGIN_BRANCH = 5,

  /**
   * @generated from enum value: GIT_FORK_SYNC_BLOCK_REASON_DIVERGED = 6;
   */
  DIVERGED = 6,
}